requirements.txt
amqp==5.2.0
asgiref==3.8.1
autopep8==2.0.4
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorclass==2.2.2
cryptography==43.0.3
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
django-redis==5.4.0
django-sslserver==0.22
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-jsonapi==7.0.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-yasg==1.21.7
google==3.0.0
google-api-core==2.21.0
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.27.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
load-dotenv==0.1.0
msgpack==1.1.0
mysqlclient==2.2.4
numpy==2.1.3
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pip-upgrader==1.4.15
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.11.1
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.2.0
requests==2.32.3
routers==0.10.1
rsa==4.9
setuptools==70.3.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
terminaltables==3.1.10
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
wheel==0.44.0
